Implement slow_start feature for mirror sync status
the slow_start parameter in the mirror class will keep the mirror
status as never discovered(-1) on the first #slow_start changes.
this will prevent falsely behavior when restarting the application.

next patch will define a rule which will cause the MirrorAPI to
always serve sites which weren't discovered yet.